Move metadata to feature files
